# Conflicts:
#	aiurt-base/aiurt-base-core/src/main/java/com/aiurt/boot/constant/SysParamCodeConstant.java
#	aiurt-module-fault/aiurt-module-fault-biz/src/main/java/com/aiurt/modules/common/api/DailyFaultApiImpl.java
#	aiurt-module-fault/aiurt-module-fault-biz/src/main/java/com/aiurt/modules/fault/service/impl/FaultServiceImpl.java
#	aiurt-module-patrol/aiurt-module-patrol-biz/src/main/java/com/aiurt/boot/api/PatrolApiServiceImpl.java
#	aiurt-module-patrol/aiurt-module-patrol-biz/src/main/java/com/aiurt/boot/task/mapper/xml/PatrolTaskMapper.xml
#	aiurt-module-system/src/main/java/com/aiurt/modules/subsystem/service/impl/CsSubsystemServiceImpl.java
#	aiurt-module-system/src/main/java/com/aiurt/modules/system/mapper/xml/CsUserSubsystemMapper.xml

+++ aiurt-module-fault/aiurt-module-fault-biz/src/main/java/com/aiurt/modules/largescream/mapper/FaultInformationMapper.java
@@ -332,7 +332,7 @@ public interface FaultInformationMapper {
      * @param endTime
      * @return
      */
-    Long getUserTimes(@Param("userId") String userId,@Param("startTime") String startTime,@Param("endTime") String endTime);
+    Integer getUserTimes(@Param("userId") String userId,@Param("startTime") String startTime,@Param("endTime") String endTime);
 
     /**
      * 查询施工次数
@@ -428,5 +428,5 @@ public interface FaultInformationMapper {
      * @param endTime
      * @return
      */
-    Long getFilterUserTimes(@Param("userId") String userId,@Param("startTime") String startTime,@Param("endTime") String endTime);
+    Integer getFilterUserTimes(@Param("userId") String userId,@Param("startTime") String startTime,@Param("endTime") String endTime);
 }
