+++ aiurt-module-inspection/aiurt-module-inspection-api/src/main/java/com/aiurt/boot/index/dto/TeamUserDTO.java
@@ -38,10 +38,10 @@ public class TeamUserDTO {
     private String workingYears;
 
     @ApiModelProperty("检修总工时")
-    private BigDecimal inspecitonTotalTime;
-
+    private Integer inspecitonTotalTime;
+    /**巡检总工时 2023-06-12 通信6期 改为 单位秒*/
     @ApiModelProperty("巡检总工时")
-    private BigDecimal patrolTotalTime;
+    private Integer patrolTotalTime;
 
     @ApiModelProperty("维修总工时")
     private BigDecimal faultTotalTime;
